Add `markdown` parser and custom AST nodes
This is the initial implementation for an intermediate layer between the
basalt TUI application and pulldown-cmark. This enables segregation of
the parsing logic.

Essentially I don't want to leak implementation details of
pulldown-cmark to the basalt TUI application.

This initial implementation does not cover all of the possible markdown
events or tags, and these are intentionally left out for now. i.e. these
won't be in the AST node tree, thus impossible to render when using this
intermediate layer.

Ordered list items do not work yet, although there exists some code for
that logic specifically.

The implemented `Parser` processes pulldown-cmark `Event`s, eventually
building up the output of AST nodes.

This implementation expects a stringified form of markdown input that is
given to either `basalt_core::markdown::from_str` function or to the
`parse` function from the `basalt_core::markdown::Parser`.

The `from_str` is a convenience function that runs `Parser::new` and
`Parser::parse` for the provided string.

Changelog: added
Module: basalt-core
Scope: markdown