Fix skipping text nodes in markdown parser
Previously, quite naively every 'unhandled' tag was pushed as a new
paragraph node, however, this did not work as expected and actually
might replace the current node, even, if that wasn't the intention.

This happened for example with styled text using emphasis or bold
markers.

This is now fixed and the tags that do not have implementation are not
handled at all as separate nodes, but instead if they contain text. This
text will be pushed to the current node block instead.

New more comprehensive test was also added and replaces the simple block
quote test.

Module: basalt-core
Scope: markdown