Add `source_range` field to Nodes
The source range is going to be necessary in the future to enable block
level editing in basalt. The source range will be used to get a slice
from the original source text.

The pulldown-cmark library had a solution for this called
`TextMergeWithOffset` and `OffsetIter`. Essentially this will return the
Event and the source range when iterating over the pulldown-cmark
events.

Additionally the `Node` was refactored into a struct instead of enum and
the previous `Node` enum was renamed to `MarkdownNode`. This was due to
all of the nodes having source_range property. This is a breaking
change.

Use `basalt-core` from crates to allow basalt-widgets to still work
locally.

Changelog: breaking
Module: basalt-core
Scope: markdown