Use `Path` instead `PathBuf` when loading config from path
The function can take a reference and does not need to take the
ownership of the value, which in my opinion will be more user friendly.

Otherwise the function will force a move and users will have to clone
the passed value if they wish to use it later in the body.

Changelog: breaking
Module: basalt-core
Scope: obsidian