Return `Vec<&Vault>` from `vaults()` instead of `Iterator`
Returning an iterator felt surprising and made the callers do more work
than perhaps necessary. The intention usually is to capture the vector
of all the vaults and not necessarily modify the values or call next
separately. That can be achieved by turning the `Vec` into iterator
instead, which to me is more idiomatic Rust anyway.

Changelog: breaking
Module: basalt-core
Scope: obsidian