Add `vault_selector` and `vault_selector_modal`
Vault selector is used to select a vault from a vec of given vaults.

Internally uses the ratatui List widget and ListState struct to manage
the list operations.

Vaults that are marked open by Obsidian are shown with a symbol ◆
marking that it is considered 'opened'.

Vault selector modal is used when a vault is already loaded and the user
wants to change the vault. Modal can be opened on top of existing screen
and contains the vault selector within that enables changing the current
vault.

Changelog: added
Module: basalt