Add `sidepanel` module
Sidepanel is used to navigate the notes within a vault.

It uses the ratatui List widget and ListState struct to manage and
operate the list state.

There is a custom selection index and that selection index will
indicate, which note is considered currently open. The currently open
note is displayed with a ◆ symbol.

Additionally the Sidepanel can be toggled, which will reduce the size of
the panel to 5 and indicate that the sidepanel is 'closed' and 'freeing'
more space for content. However the layout size change needs to be also
handled in the caller side.

Changelog: added
Module: basalt