Add `app` module
App module is the main basalt app. The app holds the complete state of
the application and is built with ELM architecture, which means that the
view and the state management is separated. The state updates are
controlled by events called actions and are listed in the `Action` enum.

The basalt app has two screens Start and Main.

The start screen displays the application 'Splash' screen and the Main
screen as the name states is the screen with the main functionality of
the app like note selection and viewing.

The keybindings are also setup in the app module inside the
`handle_press_key_event` function.

Changelog: added
Module: basalt