Add `VaultEntry` to obsidian module and remove old implementation
This commit contains a breaking change to the obsidian module vault API.

VaultEntry should be created from `&Path` using try_into conversion.

The VaultEntry is accompanied with a trait called FindNote, which is
implemented for a Vec<VaultEntry> and VaultEntry. FindNote trait
provides a method called `find_note`, which essentially finds a note
according to the given path if it exists.

This commit adds support for recursive reading of Obsidian vaults, which
enables to read all files and directories in the vault itself.
Previously a single depth of files was supported.

The implementation in `Vault` for getting the notes was changed to
`entries` instead, which returns a vector of vault entries from the
vault path filtering out any hidden files or directories like
`.obsidian`. The returned vector is not sorted alphabetically, but
instead the caller needs to ensure sorting if they want the list sorted
on their end.

The `DirEntry` implementation was removed as obsolete.

New error variant EmptyFilename was added, which is used when the
filename is empty for whatever reason. If we encounter this error when
recursively building the tree, we skip the failing entry and continue.

Changelog: breaking
Package: basalt-core
Module: obsidian