Refactor state management
Use modules for component update and event handling isolation and
separation. This should make the code easier to reason with. Eventually
these modules will be moved into their respective files, and maybe
abstracted with a trait or something similar, but could also be kept as
is. One thing that needs to be explored is the return type of the update
functions. I think these update functions should be able to change
state, but also to delegate additional events or commands. I believe
this is also called 'effects pattern'.

The 'Action's were renamed to 'Message's to match the terminology of
Elm Architecture more precisely. The Message modal is a collection of
'Global' messages and dedicated component messages. This ensures that
the top level message can be used ergonomically to update the complete
App state, since often one component can modify other component or the
global app state.

Additionally, I added more keybindings to help with navigation. For
example ESC can now be used to close modals, and arrow keys can be used
to navigate notes. The new keybindings are captured in the help.txt
file.

The modes (SELECT, NORMAL) were completely removed as obsolete and now
the status bar no longer shows the filename or mode, but instead it
shows the currently active pane.

Start widget was renamed to Splash widget.

Changelog: changed
Package: basalt