Add experimental note editor support
Notes can now be edited if `experimental_editor` configuration is set to
true in the users' config file. Otherwise, the note is displayed in a
similar 'Read' only mode as previously.

The editing experience tries to resemble the one in Obsidian where the
note is always completely 'rendered' except for the node directly under
the cursor, which is displayed with the raw data like markdown special
symbols `#`, `>`, `*`, etc.

New configuration commands where added for the experimental note editor.
These can be modified freely by the user. The only exception for now is
the EDIT mode key bindings, which cannot be modified in the user config.

The editor widget is a third-party solution called `tui-textarea` and is
well suited for this proof of concept and testing phase. However, for
the final version there probably needs to be a completely custom editor
widget to create the WYSIWYG experience.

Update `help.txt` file to match with the new editor feature

Changelog: added
Package: basalt