Change focus switch to support previous and next panes
Previously one could only travel from explorer to note editor and vice
versa. However, this does not work when additional panes are added to
the UI. This implementation is a simple approach to achieve next and
previous pane support in anticipation of outline feature.

The next and previous panes are controlled in respective update
handlers. As an example, if note editor's next pane would be outline.
Invoking the handler with message::SwitchNextPane would have update
logic to change to that pane.

Changelog: changed
Package: basalt